+++ src/apollo/types.d.ts
@@ -13,6 +13,16 @@ export type Scalars = {
   Upload: any;
 };
 
+export type AddMembersToOrganizationInput = {
+  organizationId: Scalars['Int'];
+  emails: Array<Scalars['String']>;
+};
+
+export enum AnswerType {
+  Scale = 'SCALE',
+  Text = 'TEXT'
+}
+
 export type AuthenticateInput = {
   providerAccessToken: Scalars['String'];
 };
@@ -22,11 +32,30 @@ export type AuthenticateOutput = {
   accessToken: Scalars['String'];
 };
 
+export type CreateOrganizationInput = {
+  name: Scalars['String'];
+  abbreviation?: Maybe<Scalars['String']>;
+  advisor?: Maybe<Scalars['String']>;
+  associatedFaculty?: Maybe<Scalars['String']>;
+  description?: Maybe<Scalars['String']>;
+  facebookPage?: Maybe<Scalars['String']>;
+  instagram?: Maybe<Scalars['String']>;
+  lineOfficialAccount?: Maybe<Scalars['String']>;
+  email?: Maybe<Scalars['String']>;
+  contactFullName?: Maybe<Scalars['String']>;
+  contactEmail?: Maybe<Scalars['String']>;
+  contactPhoneNumber?: Maybe<Scalars['String']>;
+  contactLineId?: Maybe<Scalars['String']>;
+  profilePicture?: Maybe<Scalars['Upload']>;
+};
+
 export type Event = {
   __typename?: 'Event';
   id: Scalars['Int'];
   organizationId: Scalars['Int'];
+  organization: Organization;
   locationId?: Maybe<Scalars['Int']>;
+  location?: Maybe<Location>;
   description: Scalars['String'];
   name: Scalars['String'];
   coverImageUrl?: Maybe<Scalars['String']>;
@@ -35,23 +64,76 @@ export type Event = {
   posterImageHash?: Maybe<Scalars['String']>;
   profileImageUrl?: Maybe<Scalars['String']>;
   profileImageHash?: Maybe<Scalars['String']>;
-  contact: Scalars['String'];
   attendeeLimit: Scalars['Int'];
-  organization: Organization;
-  location?: Maybe<Location>;
+  contact?: Maybe<Scalars['String']>;
+  questionGroups: Array<QuestionGroup>;
+  durations: Array<EventDuration>;
   tags: Array<Tag>;
+};
+
+export type EventDuration = {
+  __typename?: 'EventDuration';
+  id: Scalars['Int'];
+  eventId: Scalars['Int'];
+  start: Scalars['String'];
+  finish: Scalars['String'];
+  event: Event;
+};
+
+export type EventDurationInput = {
+  id: Scalars['Int'];
+  eventId: Scalars['Int'];
+  start: Scalars['String'];
+  finish: Scalars['String'];
+  event: EventInput;
+};
+
+export type EventInput = {
+  id: Scalars['Int'];
+  organizationId: Scalars['Int'];
+  organization: OrganizationInput;
+  locationId?: Maybe<Scalars['Int']>;
+  location?: Maybe<LocationInput>;
+  description: Scalars['String'];
+  name: Scalars['String'];
+  coverImageUrl?: Maybe<Scalars['String']>;
+  coverImageHash?: Maybe<Scalars['String']>;
+  posterImageUrl?: Maybe<Scalars['String']>;
+  posterImageHash?: Maybe<Scalars['String']>;
+  profileImageUrl?: Maybe<Scalars['String']>;
+  profileImageHash?: Maybe<Scalars['String']>;
+  attendeeLimit: Scalars['Int'];
+  contact?: Maybe<Scalars['String']>;
+  questionGroups: Array<QuestionGroupInput>;
+  durations: Array<EventDurationInput>;
+  tags: Array<TagInput>;
 };
 
 export type FileUploadInput = {
   file: Scalars['Upload'];
 };
 
+export enum Gender {
+  M = 'M',
+  F = 'F',
+  Ns = 'NS'
+}
+
 export type Location = {
   __typename?: 'Location';
   id: Scalars['Int'];
   name: Scalars['String'];
-  description?: Maybe<Scalars['String']>;
   googleMapUrl: Scalars['String'];
+  description?: Maybe<Scalars['String']>;
+  travelInformationImageUrl?: Maybe<Scalars['String']>;
+  travelInformationImageHash?: Maybe<Scalars['String']>;
+};
+
+export type LocationInput = {
+  id: Scalars['Int'];
+  name: Scalars['String'];
+  googleMapUrl: Scalars['String'];
+  description?: Maybe<Scalars['String']>;
   travelInformationImageUrl?: Maybe<Scalars['String']>;
   travelInformationImageHash?: Maybe<Scalars['String']>;
 };
@@ -59,7 +141,11 @@ export type Location = {
 export type Mutation = {
   __typename?: 'Mutation';
   authenticate: AuthenticateOutput;
+  signInWithServiceAccount: AuthenticateOutput;
   upload: Scalars['Boolean'];
+  createOrganization: Scalars['Boolean'];
+  addMembersToOrganization: Organization;
+  updateUser: User;
 };
 
 
@@ -68,8 +154,28 @@ export type MutationAuthenticateArgs = {
 };
 
 
+export type MutationSignInWithServiceAccountArgs = {
+  serviceAccount: Scalars['String'];
+};
+
+
 export type MutationUploadArgs = {
   input: FileUploadInput;
+};
+
+
+export type MutationCreateOrganizationArgs = {
+  input: CreateOrganizationInput;
+};
+
+
+export type MutationAddMembersToOrganizationArgs = {
+  input: AddMembersToOrganizationInput;
+};
+
+
+export type MutationUpdateUserArgs = {
+  input: UpdateUserInput;
 };
 
 export type Organization = {
@@ -91,11 +197,94 @@ export type Organization = {
   contactLineId?: Maybe<Scalars['String']>;
   profilePictureUrl?: Maybe<Scalars['String']>;
   profilePictureHash?: Maybe<Scalars['String']>;
+  events: Array<Event>;
+};
+
+export type OrganizationInput = {
+  id: Scalars['Int'];
+  name: Scalars['String'];
+  isVerified: Scalars['Boolean'];
+  abbreviation?: Maybe<Scalars['String']>;
+  advisor?: Maybe<Scalars['String']>;
+  associatedFaculty?: Maybe<Scalars['String']>;
+  description?: Maybe<Scalars['String']>;
+  facebookPage?: Maybe<Scalars['String']>;
+  instagram?: Maybe<Scalars['String']>;
+  lineOfficialAccount?: Maybe<Scalars['String']>;
+  email?: Maybe<Scalars['String']>;
+  contactFullName?: Maybe<Scalars['String']>;
+  contactEmail?: Maybe<Scalars['String']>;
+  contactPhoneNumber?: Maybe<Scalars['String']>;
+  contactLineId?: Maybe<Scalars['String']>;
+  profilePictureUrl?: Maybe<Scalars['String']>;
+  profilePictureHash?: Maybe<Scalars['String']>;
+  events: Array<EventInput>;
 };
 
 export type Query = {
   __typename?: 'Query';
   upcomingEvents: Array<Event>;
+  event: Event;
+  organizations: Array<Organization>;
+  organization: Organization;
+  currentUser: User;
+};
+
+
+export type QueryEventArgs = {
+  id: Scalars['Int'];
+};
+
+
+export type QueryOrganizationArgs = {
+  id: Scalars['Int'];
+};
+
+export type Question = {
+  __typename?: 'Question';
+  id: Scalars['Int'];
+  questionGroupId: Scalars['Int'];
+  seq: Scalars['Int'];
+  answerType: AnswerType;
+  isOptional: Scalars['Boolean'];
+  title: Scalars['String'];
+  subtitle: Scalars['String'];
+};
+
+export type QuestionGroup = {
+  __typename?: 'QuestionGroup';
+  id: Scalars['Int'];
+  eventId: Scalars['Int'];
+  event: Event;
+  type: QuestionGroupType;
+  seq: Scalars['Int'];
+  title: Scalars['String'];
+  questions: Array<Question>;
+};
+
+export type QuestionGroupInput = {
+  id: Scalars['Int'];
+  eventId: Scalars['Int'];
+  event: EventInput;
+  type: QuestionGroupType;
+  seq: Scalars['Int'];
+  title: Scalars['String'];
+  questions: Array<QuestionInput>;
+};
+
+export enum QuestionGroupType {
+  PreEvent = 'PRE_EVENT',
+  PostEvent = 'POST_EVENT'
+}
+
+export type QuestionInput = {
+  id: Scalars['Int'];
+  questionGroupId: Scalars['Int'];
+  seq: Scalars['Int'];
+  answerType: AnswerType;
+  isOptional: Scalars['Boolean'];
+  title: Scalars['String'];
+  subtitle: Scalars['String'];
 };
 
 export type Tag = {
@@ -104,3 +293,47 @@ export type Tag = {
   name: Scalars['String'];
 };
 
+export type TagInput = {
+  id: Scalars['Int'];
+  name: Scalars['String'];
+};
+
+export type UpdateUserInput = {
+  firstName?: Maybe<Scalars['String']>;
+  lastName?: Maybe<Scalars['String']>;
+  email?: Maybe<Scalars['String']>;
+  nickname?: Maybe<Scalars['String']>;
+  address?: Maybe<Scalars['String']>;
+  gender?: Maybe<Gender>;
+  profilePicture?: Maybe<Scalars['Upload']>;
+};
+
+
+export type User = {
+  __typename?: 'User';
+  id: Scalars['Int'];
+  firstName: Scalars['String'];
+  lastName: Scalars['String'];
+  email: Scalars['String'];
+  nickname?: Maybe<Scalars['String']>;
+  chulaId?: Maybe<Scalars['String']>;
+  address?: Maybe<Scalars['String']>;
+  profilePictureUrl?: Maybe<Scalars['String']>;
+  isChulaStudent: Scalars['Boolean'];
+  didSetup: Scalars['Boolean'];
+  gender: Gender;
+};
+
+export type UserInput = {
+  id: Scalars['Int'];
+  firstName: Scalars['String'];
+  lastName: Scalars['String'];
+  email: Scalars['String'];
+  nickname?: Maybe<Scalars['String']>;
+  chulaId?: Maybe<Scalars['String']>;
+  address?: Maybe<Scalars['String']>;
+  profilePictureUrl?: Maybe<Scalars['String']>;
+  isChulaStudent: Scalars['Boolean'];
+  didSetup: Scalars['Boolean'];
+  gender: Gender;
+};
