+++ src/modules/eventRegister/type.ts
@@ -1,0 +1,4 @@
+import { Question } from "./../../apollo/types.d";
+export interface QuestionWithAnswer extends Question {
+  answer?: string;
+}
