+++ src/modules/question/question-text/QuestionText.vue
@@ -34,7 +34,6 @@ export default defineComponent({
   emits: [USER_INPUT],
   setup(_, context) {
     const { userAnswer, userChange } = useQuestionText(context);
-
     return { userAnswer, userChange };
   }
 });
